+++ classes/Mapper.class.php
@@ -19,6 +19,11 @@ namespace Locator;
 */
 class Mapper
 {
+    protected $is_mapper = false;
+    protected $is_geocoder = false;
+    protected $display_name = 'Undefined';
+    protected $name = 'Undefined';
+
     public static function getInstance($name='')
     {
         global $_CONF_GEO;
@@ -34,6 +39,34 @@ class Mapper
             }
         }
         return $mappers[$name];
+    }
+
+
+    public static function getMapper()
+    {
+        global $_CONF_GEO;
+
+        return self::getInstance($_CONF_GEO['mapper']);
+    }
+
+
+    public static function getGeocoder()
+    {
+        global $_CONF_GEO;
+
+        return self::getInstance($_CONF_GEO['geocoder']);
+    }
+
+
+    public function getDisplayName()
+    {
+        return $this->display_name;
+    }
+
+
+    public function getName()
+    {
+        return $this->name;
     }
 
 
@@ -88,6 +121,87 @@ class Mapper
             COM_errorLog('LOCATOR: Missing url_fopen and curl support');
         }
         return $result;
+    }
+
+
+    /**
+     * Get all providers into an array
+     *
+     * @return  array   Array of objects indexed by name
+     */
+    public static function getAll()
+    {
+        static $A = NULL;
+
+        if ($A === NULL) {
+            $files = glob(__DIR__ . '/Mappers/*.class.php');
+            foreach ($files as $file) {
+                $tmp = pathinfo($file, PATHINFO_FILENAME);
+                $tmp = explode('.', $tmp);
+                $cls = '\\Locator\\Mappers\\' . $tmp[0];
+                $M = self::getInstance($tmp[0]);
+                $A[$M->getName()] = $M;
+            }
+        }
+        return $A;
+    }
+
+
+    /**
+     * Get an array of all Geocoding providers
+     *
+     * @return  array   Array of objects indexed by name
+     */
+    public static function getGeocoders()
+    {
+        $mappers = self::getAll();
+        $A = array();
+        foreach ($mappers as $name=>$mapper) {
+            if ($mapper->isGeocoder()) {
+                $A[$name] = $mapper;
+            }
+        }
+        return $A;
+    }
+
+
+    /**
+     * Get an array of all Mapping providers.
+     *
+     * @return  array   Array of objects indexed by name
+     */
+    public static function getMappers()
+    {
+        $mappers = self::getAll();
+        $A = array();
+        foreach ($mappers as $name=>$mapper) {
+            if ($mapper->isMapper()) {
+                $A[$name] = $mapper;
+            }
+        }
+        return $A;
+    }
+
+
+    /**
+     * Check if this provider is a Mapping provider
+     *
+     * @return  boolean     True or False
+     */
+    public function isMapper()
+    {
+        return $this->is_mapper;
+    }
+
+
+    /**
+     * Check if this provider is a Geocoding provider
+     *
+     * @return  boolean     True or False
+     */
+    public function isGeocoder()
+    {
+        return $this->is_geocoder;
     }
 
 }   // class Mapper