+++ language/german_formal_utf-8.php
@@ -134,6 +134,7 @@ $LANG_confignames['locator'] = array(
     'default_permissions' => 'Standard-Berechtigungen',
     'mapper'            => 'Map Provider',
     'mapquest_key'      => 'Mapquest Client Key',
+    'geocodio_api_key'  => 'Geocodio API Key',
 );
 
 /**