+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # glLocator - Change Log
 
 ## 1.2.0 (Unreleased)
-- Add Mapquest map provider
 - Add separate Google API key for Javascript
+- Add Mapquest, OpenStreetMap map providers
+- Add U.S. Census and Geocodio geocoding services
 - Fix output var in service_getCoords function, caused 500 errors
 - Add cache support for glFusion 1.8.0+